requirements.txt
amqp==5.1.1
anyio==3.7.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
channels==4.0.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.2.7
cron-descriptor==1.4.0
Django==4.2.1
django-jet-reboot==1.3.3
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
exceptiongroup==1.1.2
h11==0.14.0
httptools==0.6.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
kombu==5.3.1
packaging==23.1
Pillow==9.5.0
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg==3.1.10
psycopg2-binary==2.9.7
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
redis==4.6.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
uvicorn==0.23.2
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
